requirements.txt
anthropic>=0.49.0
python-dotenv>=1.1.0
pydantic>=2.11.3
python-docx>=1.1.2
pypdf>=5.4.0
fastapi>=0.115.0
uvicorn>=0.34.0
streamlit>=1.44.0
python-multipart>=0.0.20
sqlalchemy>=2.0.0
googlesearch-python>=1.2.3
beautifulsoup4>=4.12.0
aiohttp>=3.9.0 